Move common methods across all plugin classes to abstract base class
