+++ foo-bar.php
@@ -29,6 +29,7 @@
  */
 
 if ( version_compare( phpversion(), '5.3', '>=' ) ) {
+	require_once __DIR__ . '/php/class-plugin-base.php';
 	require_once __DIR__ . '/php/class-plugin.php';
 	$class_name = '\FooBar\Plugin';
 	$GLOBALS['foo_bar_plugin'] = new $class_name();
